More Carefully Selecting Dimensions for Unit Tests
If we choose two time dimensions, it may be that we get problematic
situations with special cases. One example is the burma14 case of the
tEHBSA256d data from the TSP Suite Example. Here, we have runs which
do not intersect in the time range, i.e., one run whose first point
appears after the last point of another run in time. This causes
aggregation to go haywire, if we plot another time dimension (y) over
a time dimension (x), as we may have -INFINITY and +INFINITY in the
data due to default point values. Something like that. Hope this is
fixed now.